I abstract more new code out of the share multicall hook into standalone libs.
