+++ hooks/useTBAPurchase.tsx
@@ -1,13 +1,11 @@
 import { BigNumber } from "ethers"
 import { useState } from "react"
-import { sepolia } from "viem/chains"
 import multicallAbi from "@/lib/abi/multicall3.json"
 import getMintMulticallCalls from "@/lib/getMintMulticallCalls"
 import {
   CATALOG_PRICE,
   CHAIN_ID,
   MULTICALL3_ADDRESS,
-  SOUND_SEPOLIA_DROP_ADDRESS,
   SPOTIFY_DROP_ADDRESS,
   ZORA_PRICE,
 } from "@/lib/consts"
@@ -17,8 +15,8 @@ import getZoraNextTokenId from "@/lib/getZoraNextTokenId"
 import getAccount from "@/lib/tokenbound/getAccount"
 import getCosignMintCall from "@/lib/getCosignMintCall"
 import getTBAInitializeCall from "@/lib/getTBAInitializeCall"
-import getSoundBridgeTx from "@/lib/getSoundBridgeTx"
 import getSoundMintCall from "@/lib/sound/getSoundMintCall"
+import getSoundInterstellarMintCall from "@/lib/sound/getSoundInterstellarMintCall"
 import useConnectedWallet from "./useConnectedWallet"
 import usePrivySendTransaction from "./usePrivySendTransaction"
 import useWalletTransaction from "./useWalletTransaction"
@@ -50,42 +48,35 @@ const useTBAPurchase = () => {
       const tba = getAccount(zoraNextTokenId)
       const tbaInitializationCall = getTBAInitializeCall(tba)
       const soundMintCall = await getSoundMintCall(tba, CHAIN_ID, SPOTIFY_DROP_ADDRESS)
-      const soundSepoliaMintCall = (await getSoundMintCall(
-        tba,
-        sepolia.id,
-        SOUND_SEPOLIA_DROP_ADDRESS,
-      )) as any
 
       if (!soundMintCall) {
         setLoading(false)
         return false
       }
+      const interstellarMintCall = await getSoundInterstellarMintCall(
+        tba,
+        connectedWallet as `0x${string}`,
+      )
 
-      const bridgeCalls = await getSoundBridgeTx({
-        destinationChainId: sepolia.id,
-        originChainId: CHAIN_ID,
-        user: connectedWallet,
-        txs: [
-          {
-            to: soundSepoliaMintCall.target,
-            data: soundSepoliaMintCall.callData,
-            value: soundSepoliaMintCall.value.toString(),
-          },
-        ],
-      })
       console.log("SWEETS soundMintCall", soundMintCall)
-      console.log("SWEETS bridgeCall same as soundMintCall?", bridgeCalls)
+      console.log("SWEETS bridgeCall same as soundMintCall?", interstellarMintCall)
 
       const soundMintCallValue = BigNumber.from(soundMintCall.value)
       const cosignMintCall = getCosignMintCall(tba)
       const cosignMintValue = BigNumber.from(CATALOG_PRICE)
-      const soundBridgeValue = BigNumber.from(bridgeCalls.value)
+      const soundBridgeValue = BigNumber.from(interstellarMintCall.value)
       const totalPrice = soundMintCallValue
         .add(cosignMintValue)
         .add(zoraTotalPrice)
         .add(soundBridgeValue)
       const hexValue = totalPrice.toHexString()
-      const calls = [...tbaCalls, tbaInitializationCall, soundMintCall, cosignMintCall, bridgeCalls]
+      const calls = [
+        ...tbaCalls,
+        tbaInitializationCall,
+        soundMintCall,
+        cosignMintCall,
+        interstellarMintCall,
+      ]
 
       if (isLoggedByEmail) {
         const response = await sendTxByPrivy(
